Squashed 'libbitcoinkernel-sys/bitcoin/' changes from 9cb45a10d7..d74f799650
d74f799650 Change kernel install directory temporarily
f331cbc2c4 Add BlockReader functionality to kernel library
REVERT: 9cb45a10d7 Change kernel install directory temporarily
REVERT: 02b96fedad Add BlockReader functionality to kernel library

git-subtree-dir: libbitcoinkernel-sys/bitcoin
git-subtree-split: d74f799650e74aba36e6bbc9d8409935e453b80b

+++ src/kernel/blockreader/blockreader.h
@@ -108,7 +108,7 @@ BITCOINKERNEL_API const kernel_TransactionWitness* BITCOINKERNEL_WARN_UNUSED_RES
 
 BITCOINKERNEL_API size_t BITCOINKERNEL_WARN_UNUSED_RESULT kernel_witness_get_stack_size(const kernel_TransactionWitness* witness) BITCOINKERNEL_ARG_NONNULL(1);
 
-BITCOINKERNEL_API kernel_ByteArray* BITCOINKERNEL_WARN_UNUSED_RESULT kernel_witness_get_stack_item(const kernel_TransactionWitness* witness, uint32_t index) BITCOINKERNEL_ARG_NONNULL(1);
+BITCOINKERNEL_API kernel_ByteArray* BITCOINKERNEL_WARN_UNUSED_RESULT kernel_witness_get_stack_item(const kernel_TransactionWitness* witness, size_t index) BITCOINKERNEL_ARG_NONNULL(1);
 
 BITCOINKERNEL_API bool BITCOINKERNEL_WARN_UNUSED_RESULT kernel_witness_is_null(const kernel_TransactionWitness* witness) BITCOINKERNEL_ARG_NONNULL(1);
